+++ pkgs/bazel_worker/test/test_all.dart
@@ -5,9 +5,11 @@
 import 'driver_test.dart' as driver;
 import 'message_grouper_test.dart' as message_grouper;
 import 'worker_loop_test.dart' as worker_loop;
+import 'driver_connection_test.dart' as driver_connection;
 
 void main() {
   driver.main();
   message_grouper.main();
   worker_loop.main();
+  driver_connection.main();
 }